backend/requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.30.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography==45.0.4

# Configuration
python-dotenv==1.0.0

# File Upload
python-multipart==0.0.6

# Image Processing
Pillow==11.2.1

# Monitoring & Logging
redis==5.3.0

# System Dependencies
typing_extensions==4.13.2

attrs==25.3.0

distro==1.9.0

PySocks==1.7.1

protobuf==5.29.5

pyOpenSSL==25.1.0

pg8000
requests==2.32.3

pydantic[email]